篮球附属服务 - 新增实时获取赛事结果

+++ src/main/java/com/example/selenium/bo/result/ResultBodyDetails.java
@@ -53,7 +53,7 @@ public class ResultBodyDetails {
     /***
      * 是否过关
      */
-    private BigDecimal oc;
+    private Integer oc;
 
     /***
      * Details
